Add sd_check function to pyportal
Both __init__ and fetch access the in self._sdcard parameter to determine
if a SD card is installed. __init__ handles the mounting of the card.
fetch() alters the filename path as needed prior to calling wget(). This
new function will allow the calling of pyportal.sd_check() to determine
if a card is present and alter operations as needed.
Simple example:
if pyportal.sdcheck():
    print("SD card is installed")
else:
    print("SD card is NOT installed")